Update JP_2499_DRT.dctl
v1.2.1 adds an alternative option for HDR formation, by default it will behave just like it did before this update but the user can choose to "unroll highlights" this is closer to the behavior you would see in other Picture formation option. In my opinion unrolling is not the best way to do this as it re-forms the image even if slightly, but I want to give the user the freedom to choose it themselves.